+++ src/main/python/server.py
@@ -11,7 +11,10 @@ import os
 SERVER_ID = str(uuid.uuid1())
 
 LOG_NAME = "server.log"
-os.remove(LOG_NAME)
+
+if os.path.isfile(LOG_NAME):
+    os.remove(LOG_NAME)
+
 
 LOGGER = logging.getLogger(LOG_NAME)
 logging.basicConfig(filename=LOG_NAME,level=logging.DEBUG)
@@ -50,7 +53,8 @@ def error_msg(txt):
     msg = json.dumps({
         "sourceID": SERVER_ID,
         "messageID": 114,
-        "payload": txt
+        "payload": {'error': txt}
+
     })
     return msg
 
@@ -78,7 +82,8 @@ def handle_101(msg):
     sessID = CTRL.new_session()
     if CTRL.client_join(src_client, sessID):
         LOGGER.info("Client " + src_client + " joined session " + str(sessID))
-        msg = make_msg(SERVER_ID, 102, CTRL.get_session(sessID))
+        msg = make_msg(SERVER_ID, 102, {'session': CTRL.get_session(sessID)})
+
     else:
         LOGGER.error("Client " + src_client + " attempt to join session failed")
         msg = error_msg("Error: Could not join session.")
@@ -94,7 +99,8 @@ def handle_112(msg):
     LOGGER.info("handle_112():Client Connect started")
 
     # No check for sourceID in this function b/c a new Client will not yet have one
-    nick = msg.get("payload")
+    nick = msg.get("payload").get('nickname')
+
     if not nick:
         LOGGER.error("Client did not provide nickname")
         return error_msg("Error: Nickname not provided")
@@ -102,7 +108,8 @@ def handle_112(msg):
     clientID = CTRL.new_client(nick)
     LOGGER.info("New client ID: " + clientID)
 
-    return make_msg(SERVER_ID, 113, clientID)
+    return make_msg(SERVER_ID, 113, {'clientID':clientID})
+
